### 2024-01-18 #6239 @editorialbot
[PRE REVIEW]: SitCoM: Modular framework for the integration of (extreme) events and their impacts in Unity
**Submitting author:** @ruedi99ms (<a href="http://orcid.org/0009-0006-4168-1115">Niklas Suhre</a>)
**Repository:** https://github.com/ruedi99ms/SitCoM
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
