### 2024-02-17 #6239 @osorensen
👋 @shahchiragh, @mikemahoney218, @ApocalyVec, @ahmadawais, @luxaritas, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
