### 2024-02-17 #6239 @mikemahoney218
Sorry, I don't currently have the bandwidth (or frankly, the C# experience) to give this a good review.
