### 2024-04-01 #6458 @luxaritas
Thanks for the nudge - just took an initial look through this.

Before I go any further, I want to raise a concern around JOSS's "substantial scholarly effort" requirement.

From my understanding, the work presented here is principally in [https://github.com/ruedi99ms/SitCoM/tree/main/Assets/_SitCoM](Assets/_SitCoM), which includes:
* An GUI that presents a list of available "scenarios" and "impacts" that can be selected in addition to timing
* Base classes/manager classes that are set up to call user-defined "scenario" and "impact" functions when they are selected at the GUI, and time falls within the specified interval
* A GUI for providing names and descriptions of scenarios and impacts (which also feels unnecessary, as opposed to using a text file or global variable/enum somewhere?)

I appreciate the value proposition here in being able to create "plug and play" simulations for visualizing extreme events and their impacts. However it doesn't seem to me that the scope as-implemented matches the size JOSS requires. Being mostly generating a UI of switches which conditionally enable objects, this feels like it falls under "minor utility package" territory and does not appear to me to be something that would take >= 3 months of time in the context of spinning up a new unity project that needs some UI to do something similar.
