### 2024-04-02 #6458 @osorensen
Thanks a lot for your feedback @luxaritas.

@ruedi99ms, @luxaritas and @ApocalyVec, I pause the review for now, until the JOSS editorial team has evaluated whether the submission is within scope.
