### 2024-05-05 #6458 @oliviaguest
Dear author(s), @ruedi99ms, I am sorry to say that JOSS requires certain conditions to be met for this to remain under review, such as the requiremnt to meet our idea for [Substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), which includes but is not limited to the following (see more at previous link):

> As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual. Some factors that may be considered by editors and reviewers when judging effort include:
> * Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.

I am sorry and I hope this does not deter you from choosing JOSS in the future.

Best,
Olivia
